# frontend/src/api/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Ground Truth Curator",
		"version": "1.0.6"
	},
	"paths": {
...
		"/v1/ground-truths": {
			"post": {
				"tags": ["ground-truths"],
				"summary": "Import Bulk",
				"description": "Bulk import ground truth items.\n\nBehavior:\n- ID field: Each item may include 'id' (alias 'uuid'); missing/blank values are generated server-side using\n  randomname to produce a two-word, hyphenated string (e.g., 'sleek-voxel').\n- Order preservation: Response `uuids` mirrors request order for easy correlation.\n- Duplicate handling: Existing items are not overwritten; per-item errors are returned in `errors`.\n- Optional approval: `approve=true` marks all items approved and sets review metadata.",
				"operationId": "import_bulk_v1_ground_truths_post",
				"parameters": [
					{
						"name": "buckets",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "integer",
									"maximum": 50,
									"minimum": 1
								},
								{
									"type": "null"
								}
							],
							"title": "Buckets"
						}
					},
					{
						"name": "approve",
						"in": "query",
						"required": false,
						"schema": {
							"type": "boolean",
							"description": "If true, mark all imported items as approved and set review metadata.",
							"default": false,
							"title": "Approve"
						},
						"description": "If true, mark all imported items as approved and set review metadata."
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/GroundTruthItem-Input"
								},
								"title": "Items"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ImportBulkResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"get": {
				"tags": ["ground-truths"],
				"summary": "List All Ground Truths",
				"operationId": "list_all_ground_truths_v1_ground_truths_get",
				"parameters": [
					{
						"name": "status",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Status"
						}
					},
					{
						"name": "dataset",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Dataset"
						}
					},
					{
						"name": "tags",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "Tags"
						}
					},
					{
						"name": "itemId",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Search for items by ID (case-sensitive partial match)",
							"title": "Itemid"
						},
						"description": "Search for items by ID (case-sensitive partial match)"
					},
					{
						"name": "refUrl",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Search for items by reference URL (case-sensitive partial match)",
							"title": "Refurl"
						},
						"description": "Search for items by reference URL (case-sensitive partial match)"
					},
					{
						"name": "sortBy",
						"in": "query",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortField",
							"default": "reviewedAt"
						}
					},
					{
						"name": "sortOrder",
						"in": "query",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortOrder",
							"default": "desc"
						}
					},
					{
						"name": "page",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"default": 1,
							"title": "Page"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"default": 25,
							"title": "Limit"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroundTruthListResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
		"/v1/ground-truths/{datasetName}": {
			"get": {
				"tags": ["ground-truths"],
				"summary": "List Ground Truths",
				"operationId": "list_ground_truths_v1_ground_truths__datasetName__get",
				"parameters": [
					{
						"name": "datasetName",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Datasetname"
						}
					},
					{
						"name": "status",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/GroundTruthStatus"
								},
								{
									"type": "null"
								}
							],
							"title": "Status"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/GroundTruthItem-Output"
									},
									"title": "Response List Ground Truths V1 Ground Truths  Datasetname  Get"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/v1/ground-truths/{datasetName}/{bucket}/{item_id}": {
			"get": {
				"tags": ["ground-truths"],
				"summary": "Get Ground Truth",
				"operationId": "get_ground_truth_v1_ground_truths__datasetName___bucket___item_id__get",
				"parameters": [
					{
						"name": "datasetName",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Datasetname"
						}
					},
					{
						"name": "bucket",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Bucket"
						}
					},
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Item Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroundTruthItem-Output"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["ground-truths"],
				"summary": "Update Ground Truth",
				"operationId": "update_ground_truth_v1_ground_truths__datasetName___bucket___item_id__put",
				"parameters": [
					{
						"name": "datasetName",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Datasetname"
						}
					},
					{
						"name": "bucket",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Bucket"
						}
					},
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Item Id"
						}
					},
					{
						"name": "If-Match",
						"in": "header",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "If-Match"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"additionalProperties": true,
								"title": "Payload"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroundTruthItem-Output"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["ground-truths"],
				"summary": "Delete Item",
				"description": "Soft-delete a ground truth item.",
				"operationId": "delete_item_v1_ground_truths__datasetName___bucket___item_id__delete",
				"parameters": [
					{
						"name": "datasetName",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Datasetname"
						}
					},
					{
						"name": "bucket",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"title": "Bucket"
						}
					},
					{
						"name": "item_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Item Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": {
										"type": "string"
									},
									"title": "Response Delete Item V1 Ground Truths  Datasetname   Bucket   Item Id  Delete"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/v1/ground-truths/recompute-tags": {
			"post": {
				"tags": ["ground-truths"],
				"summary": "Recompute Computed Tags",
				"description": "Recompute computed tags for all matching ground truth items.\n\nThis endpoint recalculates the `computedTags` field for all items matching\nthe optional filter criteria. Use this when:\n- A new computed tag plugin has been added\n- An existing plugin's logic has been modified\n- Data needs to be reprocessed after a bug fix\n\nThe operation is idempotent - running it multiple times with the same\ndata produces the same result.\n\n**Important**: By default, `dry_run=true` for safety. Set `dry_run=false`\nto actually persist changes.",
				"operationId": "recompute_computed_tags_v1_ground_truths_recompute_tags_post",
				"parameters": [
					{
						"name": "dataset",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter to items in a specific dataset. If not provided, processes all datasets.",
							"title": "Dataset"
						},
						"description": "Filter to items in a specific dataset. If not provided, processes all datasets."
					},
					{
						"name": "status",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/GroundTruthStatus"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter to items with a specific status (draft, approved, etc.).",
							"title": "Status"
						},
						"description": "Filter to items with a specific status (draft, approved, etc.)."
					},
					{
						"name": "dry_run",
						"in": "query",
						"required": false,
						"schema": {
							"type": "boolean",
							"description": "If true, preview changes without persisting. Defaults to true for safety.",
							"default": true,
							"title": "Dry Run"
						},
						"description": "If true, preview changes without persisting. Defaults to true for safety."
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecomputeTagsResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"GroundTruthListResponse": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/GroundTruthItem-Output"
						},
						"type": "array",
						"title": "Items"
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMetadata"
					}
				},
				"type": "object",
				"required": ["items", "pagination"],
				"title": "GroundTruthListResponse"
			},
			"GroundTruthStatus": {
				"type": "string",
				"enum": ["draft", "approved", "deleted", "skipped"],
				"title": "GroundTruthStatus"
			},
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"ImportBulkResponse": {
				"properties": {
					"imported": {
						"type": "integer",
						"title": "Imported",
						"description": "Number of items successfully imported."
					},
					"errors": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Errors",
						"description": "List of error messages for items that failed to import."
					},
					"uuids": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Uuids",
						"description": "The item IDs in the same order as the request payload. Clients may send either 'id' or 'uuid' in each item; when missing or blank, the server generates a two-word, hyphenated string ID (e.g., 'sleek-voxel') via randomname."
					}
				},
				"type": "object",
				"required": ["imported", "errors"],
				"title": "ImportBulkResponse"
			},
...
			"RecomputeTagsResponse": {
				"properties": {
					"total": {
						"type": "integer",
						"title": "Total",
						"description": "Total items matching the filter criteria."
					},
					"processed": {
						"type": "integer",
						"title": "Processed",
						"description": "Number of items successfully processed."
					},
					"updated": {
						"type": "integer",
						"title": "Updated",
						"description": "Number of items with changed computed tags."
					},
					"skipped": {
						"type": "integer",
						"title": "Skipped",
						"description": "Number of items with unchanged computed tags."
					},
					"errors": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Errors",
						"description": "Error messages for failed items."
					},
					"dry_run": {
						"type": "boolean",
						"title": "Dry Run",
						"description": "Whether this was a dry run (no changes persisted)."
					},
					"duration_ms": {
						"type": "integer",
						"title": "Duration Ms",
						"description": "Operation duration in milliseconds."
					}
				},
				"type": "object",
				"required": [
					"total",
					"processed",
					"updated",
					"skipped",
					"dry_run",
					"duration_ms"
				],
				"title": "RecomputeTagsResponse",
				"description": "Response for bulk computed tag recomputation."
			},
...
			"SortField": {
				"type": "string",
				"enum": [
					"reviewedAt",
					"updatedAt",
					"id",
					"hasAnswer",
					"totalReferences"
				],
				"title": "SortField"
			},
			"SortOrder": {
				"type": "string",
				"enum": ["asc", "desc"],
				"title": "SortOrder"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			}
		}
	}
}